style: change ColumnLimit to 80 in .clang-format style: add AVX2 flag to CMAKE_CXX_FLAGS in CMakeLists.txt refactor: rename MedianBlur.h to median_blur.h and update include statements refactor: move MedianBlur.cpp implementation to median_blur.h and remove MedianBlur.h The ColumnLimit in .clang-format has been changed to 80 to improve readability. The AVX2 flag has been added to CMAKE_CXX_FLAGS in CMakeLists.txt to enable AVX2 instructions. The MedianBlur.h header file has been renamed to median_blur.h to follow the naming convention. The include statements have been updated accordingly. The implementation of MedianBlur has been moved to median_blur.h and the MedianBlur.h header file has been removed.
feat(median_blur.cpp): add median blur implementation with SIMD optimization
Add a new implementation of median blur with SIMD optimization. The implementation uses TBB to parallelize the computation of the median blur. The implementation also includes a function to optimize the computation of the minimum and maximum values of an array using SIMD instructions.